refactor: Start moving NS related logic to own mod
